Split addon model and use default_factory
