+++ swiggy_order_hooks/model/cart.py
@@ -1,4 +1,4 @@
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from dataclasses_json import dataclass_json
 from typing import Optional, List
 
@@ -7,5 +7,5 @@ from .item import Item
 @dataclass_json
 @dataclass
 class Cart:
-    charges: Optional[dict] = dict
-    items: Optional[List[Item]] = list
+    charges: Optional[dict] = field(default_factory=dict)
+    items: Optional[List[Item]] = field(default_factory=list)
