+++ swiggy_order_hooks/model/restaurant_details.py
@@ -1,5 +1,5 @@
 from typing import List, Optional
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from dataclasses_json import dataclass_json
 
 @dataclass_json
@@ -7,4 +7,4 @@ from dataclasses_json import dataclass_json
 class RestaurantDetails:
     restaurant_lat: Optional[str] = None
     restaurant_lng: Optional[str] = None
-    categories: Optional[List[str]] = list 
+    categories: Optional[List[str]] = field(default_factory=list)
